+++ spi/core-spi/src/main/java/org/eclipse/dataspaceconnector/spi/event/transferprocess/TransferProcessCancelled.java
@@ -37,8 +37,8 @@ public class TransferProcessCancelled extends Event<TransferProcessCancelled.Pay
             super(new TransferProcessCancelled(), new Payload());
         }
 
-        public Builder transferProcessId(String contractDefinitionId) {
-            event.payload.transferProcessId = contractDefinitionId;
+        public Builder transferProcessId(String transferProcessId) {
+            event.payload.transferProcessId = transferProcessId;
             return this;
         }